Add basic room select modal
Add a room select modal that appears when the search box is focussed.
This then navigates to the route for the room. We haven't actually added
the logic for what to do when you navigate to a room yet, and the search
field doesn't filter the rooms at all. But it gives us an idea of what
the mobile experience might be like.