Convert config to typescript file
This lets us add comments, and use multiline strings for the
descriptions.

This does mean we need to dedent the descriptions, otherwise markdown
thinks it is a code block. We use the first non-empty line to know how
much to dedent by. This does have the disadvantage that you can't use
indentation to make the first line a code block. You can still use code
fences (```) to do the same thing though.