Move config docs to type definition
This makes the docs a bit cleaner, at the expense of removing the
colocation of the docs from the config. But that never really worked for
the rooms anyway, as they were an array.

In addition, rename the file to `config.types.ts`, which follows a more
common naming convention.