Add i18n for UI text
Use i18n for the text in the UI using the next-intl library. We still
only support a single language, but we will at least have a single place
where all the strings to translate are, and once someone has done it for
one language, they can contribute it back and it can be used by other
people.