Add /about page
This is primarily as a place to put the attribution for the favicon, but
also serves as a place to explain what the map is for, link to the con
website, things like that.

The "Back to map" link goes to `/`, rather than the room they were
looking at. This does mean they lose some context, but I've not been
able to figure out how to reliably go back if they came from the map,
but go to `/` if they were linked directly to the about page.