Add ability to focus on the selected room
When selecting a room in the search results, it now pans and zooms
to focus on the selected room.

This also adds an icon to the left of the room title in the description
bar at the bottom to allow the user to pan and zoom on the selected
room.