Add support for multiple maps
The config `map` property is now a `maps` property, and each has an id
and label. If there is more than one map, it will add a map select
component that allows the user to select the map.

The zoom and pan settings are now persisted per-map, which hopefully
makes it easier to switch back and forth between maps to understand how
to get from one room to another.

The urls now have the map in them. This means upgrading from an old map
will break all the old URLs. However, people don't tend to upgrade old
maps, so that's not really a big deal. Besides, the rooms will likely
have to be changed as a part of the upgrade anyway.